4. Image Classification APP/requirements.txt
tensorflow
keras
Pillow
Flask
numpy